+++ quickwit/quickwit-control-plane/src/scheduler.rs
@@ -24,14 +24,9 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use anyhow::Context;
-use async_trait::async_trait;
 use itertools::Itertools;
-use quickwit_actors::{Actor, ActorContext, ActorExitStatus, Handler};
 use quickwit_config::{SourceConfig, INGEST_SOURCE_ID};
 use quickwit_metastore::{ListIndexesQuery, Metastore};
-use quickwit_proto::control_plane::{
-    ControlPlaneResult, NotifyIndexChangeRequest, NotifyIndexChangeResponse,
-};
 use quickwit_proto::indexing::{ApplyIndexingPlanRequest, IndexingService, IndexingTask};
 use quickwit_proto::metastore::{ListShardsRequest, ListShardsSubrequest};
 use quickwit_proto::{NodeId, ShardId};
@@ -43,28 +38,12 @@ use crate::indexing_plan::{
 };
 use crate::{IndexerNodeInfo, IndexerPool};
 
-/// Interval between two controls (or checks) of the desired plan VS running plan.
-const CONTROL_PLAN_LOOP_INTERVAL: Duration = if cfg!(any(test, feature = "testsuite")) {
-    Duration::from_millis(500)
-} else {
-    Duration::from_secs(3)
-};
-
-/// Interval between two scheduling of indexing plans. No need to be faster than the
-/// control plan loop.
-// Note: it's currently not possible to define a const duration with
-// `CONTROL_PLAN_LOOP_INTERVAL * number`.
-const REFRESH_PLAN_LOOP_INTERVAL: Duration = if cfg!(any(test, feature = "testsuite")) {
-    Duration::from_secs(3)
-} else {
-    Duration::from_secs(60)
-};
-
-const MIN_DURATION_BETWEEN_SCHEDULING: Duration = if cfg!(any(test, feature = "testsuite")) {
-    Duration::from_millis(50)
-} else {
-    Duration::from_secs(30)
-};
+pub(crate) const MIN_DURATION_BETWEEN_SCHEDULING: Duration =
+    if cfg!(any(test, feature = "testsuite")) {
+        Duration::from_millis(50)
+    } else {
+        Duration::from_secs(30)
+    };
 
 #[derive(Debug, Clone, Default, Serialize)]
 pub struct IndexingSchedulerState {
@@ -131,26 +110,6 @@ impl fmt::Debug for IndexingScheduler {
     }
 }
 
-#[async_trait]
-impl Actor for IndexingScheduler {
-    type ObservableState = IndexingSchedulerState;
-
-    fn observable_state(&self) -> Self::ObservableState {
-        self.state.clone()
-    }
-
-    fn name(&self) -> String {
-        "IndexingScheduler".to_string()
-    }
-
-    async fn initialize(&mut self, ctx: &ActorContext<Self>) -> Result<(), ActorExitStatus> {
-        self.handle(RefreshPlanLoop, ctx).await?;
-        ctx.schedule_self_msg(CONTROL_PLAN_LOOP_INTERVAL, ControlPlanLoop)
-            .await;
-        Ok(())
-    }
-}
-
 impl IndexingScheduler {
     pub fn new(
         cluster_id: String,
@@ -167,7 +126,11 @@ impl IndexingScheduler {
         }
     }
 
-    async fn schedule_indexing_plan_if_needed(&mut self) -> anyhow::Result<()> {
+    pub fn observable_state(&self) -> IndexingSchedulerState {
+        self.state.clone()
+    }
+
+    pub(crate) async fn schedule_indexing_plan_if_needed(&mut self) -> anyhow::Result<()> {
         let mut indexers = self.get_indexers_from_indexer_pool().await;
         if indexers.is_empty() {
             warn!("No indexer available, cannot schedule an indexing plan.");
@@ -283,7 +246,7 @@ impl IndexingScheduler {
     /// chitchat cluster state. If true, do nothing.
     /// - If node IDs differ, schedule a new indexing plan.
     /// - If indexing tasks differ, apply again the last plan.
-    async fn control_running_plan(&mut self) -> anyhow::Result<()> {
+    pub(crate) async fn control_running_plan(&mut self) -> anyhow::Result<()> {
         let last_applied_plan =
             if let Some(last_applied_plan) = self.state.last_applied_physical_plan.as_ref() {
                 last_applied_plan
@@ -363,61 +326,13 @@ impl IndexingScheduler {
     }
 }
 
-#[async_trait]
-impl Handler<NotifyIndexChangeRequest> for IndexingScheduler {
-    type Reply = ControlPlaneResult<NotifyIndexChangeResponse>;
-
-    async fn handle(
-        &mut self,
-        _: NotifyIndexChangeRequest,
-        _: &ActorContext<Self>,
-    ) -> Result<Self::Reply, ActorExitStatus> {
-        debug!("Index change notification: schedule indexing plan.");
+impl IndexingScheduler {
+    // Should be called whenever a change in the list of index/shard
+    // has happened
+    pub(crate) async fn on_index_change(&mut self) -> anyhow::Result<()> {
         self.schedule_indexing_plan_if_needed()
             .await
             .context("error when scheduling indexing plan")?;
-        Ok(Ok(NotifyIndexChangeResponse {}))
-    }
-}
-
-#[derive(Debug)]
-struct ControlPlanLoop;
-
-#[async_trait]
-impl Handler<ControlPlanLoop> for IndexingScheduler {
-    type Reply = ();
-
-    async fn handle(
-        &mut self,
-        _message: ControlPlanLoop,
-        ctx: &ActorContext<Self>,
-    ) -> Result<(), ActorExitStatus> {
-        if let Err(error) = self.control_running_plan().await {
-            error!("Error when controlling the running plan: `{}`.", error);
-        }
-        ctx.schedule_self_msg(CONTROL_PLAN_LOOP_INTERVAL, ControlPlanLoop)
-            .await;
-        Ok(())
-    }
-}
-
-#[derive(Debug)]
-struct RefreshPlanLoop;
-
-#[async_trait]
-impl Handler<RefreshPlanLoop> for IndexingScheduler {
-    type Reply = ();
-
-    async fn handle(
-        &mut self,
-        _message: RefreshPlanLoop,
-        ctx: &ActorContext<Self>,
-    ) -> Result<(), ActorExitStatus> {
-        if let Err(error) = self.schedule_indexing_plan_if_needed().await {
-            error!("Error when scheduling indexing plan: `{}`.", error);
-        }
-        ctx.schedule_self_msg(REFRESH_PLAN_LOOP_INTERVAL, RefreshPlanLoop)
-            .await;
         Ok(())
     }
 }
@@ -583,373 +498,7 @@ fn get_indexing_tasks_diff<'a>(
 
 #[cfg(test)]
 mod tests {
-    use std::collections::{HashMap, HashSet};
-    use std::num::NonZeroUsize;
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    use chitchat::transport::ChannelTransport;
-    use futures::{Stream, StreamExt};
-    use quickwit_actors::{ActorHandle, Inbox, Mailbox, Universe};
-    use quickwit_cluster::{create_cluster_for_test, Cluster, ClusterChange};
-    use quickwit_common::test_utils::wait_until_predicate;
-    use quickwit_common::tower::{Change, Pool};
-    use quickwit_config::service::QuickwitService;
-    use quickwit_config::{KafkaSourceParams, SourceConfig, SourceInputFormat, SourceParams};
-    use quickwit_indexing::IndexingService;
-    use quickwit_metastore::{IndexMetadata, ListIndexesQuery, MockMetastore};
-    use quickwit_proto::indexing::{ApplyIndexingPlanRequest, IndexingServiceClient, IndexingTask};
-    use quickwit_proto::metastore::ListShardsResponse;
-    use quickwit_proto::NodeId;
-    use serde_json::json;
-
-    use super::{IndexingScheduler, CONTROL_PLAN_LOOP_INTERVAL};
-    use crate::scheduler::{
-        get_indexing_plans_diff, MIN_DURATION_BETWEEN_SCHEDULING, REFRESH_PLAN_LOOP_INTERVAL,
-    };
-    use crate::IndexerNodeInfo;
-
-    fn index_metadata_for_test(
-        index_id: &str,
-        source_id: &str,
-        desired_num_pipelines: usize,
-        max_num_pipelines_per_indexer: usize,
-    ) -> IndexMetadata {
-        let mut index_metadata = IndexMetadata::for_test(index_id, "ram://indexes/test-index");
-        let source_config = SourceConfig {
-            enabled: true,
-            source_id: source_id.to_string(),
-            max_num_pipelines_per_indexer: NonZeroUsize::new(max_num_pipelines_per_indexer)
-                .unwrap(),
-            desired_num_pipelines: NonZeroUsize::new(desired_num_pipelines).unwrap(),
-            source_params: SourceParams::Kafka(KafkaSourceParams {
-                topic: "topic".to_string(),
-                client_log_level: None,
-                client_params: json!({
-                    "bootstrap.servers": "localhost:9092",
-                }),
-                enable_backfill_mode: true,
-            }),
-            transform_config: None,
-            input_format: SourceInputFormat::Json,
-        };
-        index_metadata
-            .sources
-            .insert(source_id.to_string(), source_config);
-        index_metadata
-    }
-
-    pub fn test_indexer_change_stream(
-        cluster_change_stream: impl Stream<Item = ClusterChange> + Send + 'static,
-        indexing_clients: HashMap<String, Mailbox<IndexingService>>,
-    ) -> impl Stream<Item = Change<String, IndexerNodeInfo>> + Send + 'static {
-        cluster_change_stream.filter_map(move |cluster_change| {
-            let indexing_clients = indexing_clients.clone();
-            Box::pin(async move {
-                match cluster_change {
-                    ClusterChange::Add(node)
-                        if node.enabled_services().contains(&QuickwitService::Indexer) =>
-                    {
-                        let node_id = node.node_id().to_string();
-                        let indexing_tasks = node.indexing_tasks().to_vec();
-                        let client_mailbox = indexing_clients.get(&node_id).unwrap().clone();
-                        let client = IndexingServiceClient::from_mailbox(client_mailbox);
-                        Some(Change::Insert(
-                            node_id,
-                            IndexerNodeInfo {
-                                client,
-                                indexing_tasks,
-                            },
-                        ))
-                    }
-                    ClusterChange::Remove(node) => Some(Change::Remove(node.node_id().to_string())),
-                    _ => None,
-                }
-            })
-        })
-    }
-
-    async fn start_scheduler(
-        cluster: Cluster,
-        indexers: &[&Cluster],
-        universe: &Universe,
-    ) -> (Vec<Inbox<IndexingService>>, ActorHandle<IndexingScheduler>) {
-        let index_1 = "test-indexing-plan-1";
-        let source_1 = "source-1";
-        let index_2 = "test-indexing-plan-2";
-        let source_2 = "source-2";
-        let index_metadata_1 = index_metadata_for_test(index_1, source_1, 2, 2);
-        let mut index_metadata_2 = index_metadata_for_test(index_2, source_2, 1, 1);
-        index_metadata_2.create_timestamp = index_metadata_1.create_timestamp + 1;
-        let mut metastore = MockMetastore::default();
-        metastore.expect_list_indexes_metadatas().returning(
-            move |_list_indexes_query: ListIndexesQuery| {
-                Ok(vec![index_metadata_2.clone(), index_metadata_1.clone()])
-            },
-        );
-        metastore.expect_list_shards().returning(|_| {
-            Ok(ListShardsResponse {
-                subresponses: Vec::new(),
-            })
-        });
-        let mut indexer_inboxes = Vec::new();
-        let indexer_pool = Pool::default();
-        let change_stream = cluster.ready_nodes_change_stream().await;
-        let mut indexing_clients = HashMap::new();
-        for indexer in indexers {
-            let (indexing_service_mailbox, indexing_service_inbox) = universe.create_test_mailbox();
-            indexing_clients.insert(indexer.self_node_id().to_string(), indexing_service_mailbox);
-            indexer_inboxes.push(indexing_service_inbox);
-        }
-        let indexer_change_stream = test_indexer_change_stream(change_stream, indexing_clients);
-        indexer_pool.listen_for_changes(indexer_change_stream);
-
-        let self_node_id: NodeId = cluster.self_node_id().to_string().into();
-        let indexing_scheduler = IndexingScheduler::new(
-            cluster.cluster_id().to_string(),
-            self_node_id,
-            Arc::new(metastore),
-            indexer_pool,
-        );
-        let (_, scheduler_handler) = universe.spawn_builder().spawn(indexing_scheduler);
-        (indexer_inboxes, scheduler_handler)
-    }
-
-    #[tokio::test]
-    async fn test_scheduler_scheduling_and_control_loop_apply_plan_again() {
-        quickwit_common::setup_logging_for_tests();
-        let transport = ChannelTransport::default();
-        let cluster =
-            create_cluster_for_test(Vec::new(), &["indexer", "control_plane"], &transport, true)
-                .await
-                .unwrap();
-        cluster
-            .wait_for_ready_members(|members| members.len() == 1, Duration::from_secs(5))
-            .await
-            .unwrap();
-        let universe = Universe::with_accelerated_time();
-        let (indexing_service_inboxes, scheduler_handler) =
-            start_scheduler(cluster.clone(), &[&cluster.clone()], &universe).await;
-        let indexing_service_inbox = indexing_service_inboxes[0].clone();
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        let indexing_service_inbox_messages =
-            indexing_service_inbox.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 1);
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 1);
-        assert!(scheduler_state.last_applied_physical_plan.is_some());
-        assert_eq!(indexing_service_inbox_messages.len(), 1);
-
-        // After a CONTROL_PLAN_LOOP_INTERVAL, the control loop will check if the desired plan is
-        // running on the indexer. As chitchat state of the indexer is not updated (we did
-        // not instantiate a indexing service for that), the control loop will apply again
-        // the same plan.
-        // Check first the plan is not updated before `MIN_DURATION_BETWEEN_SCHEDULING`.
-        tokio::time::sleep(MIN_DURATION_BETWEEN_SCHEDULING.mul_f32(0.5)).await;
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 1);
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 1);
-
-        // After `MIN_DURATION_BETWEEN_SCHEDULING`, we should see a plan update.
-        tokio::time::sleep(MIN_DURATION_BETWEEN_SCHEDULING.mul_f32(0.7)).await;
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        let indexing_service_inbox_messages =
-            indexing_service_inbox.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 1);
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 2);
-        assert_eq!(indexing_service_inbox_messages.len(), 1);
-        let indexing_tasks = indexing_service_inbox_messages
-            .first()
-            .unwrap()
-            .indexing_tasks
-            .clone();
-
-        // Update the indexer state and check that the indexer does not receive any new
-        // `ApplyIndexingPlanRequest`.
-        cluster
-            .update_self_node_indexing_tasks(&indexing_tasks)
-            .await
-            .unwrap();
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 2);
-        let indexing_service_inbox_messages =
-            indexing_service_inbox.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(indexing_service_inbox_messages.len(), 0);
-
-        // Update the indexer state with a different plan and check that the indexer does now
-        // receive a new `ApplyIndexingPlanRequest`.
-        cluster
-            .update_self_node_indexing_tasks(&[indexing_tasks[0].clone()])
-            .await
-            .unwrap();
-        tokio::time::sleep(MIN_DURATION_BETWEEN_SCHEDULING.mul_f32(1.2)).await;
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 3);
-        let indexing_service_inbox_messages =
-            indexing_service_inbox.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(indexing_service_inbox_messages.len(), 1);
-        universe.assert_quit().await;
-    }
-
-    #[tokio::test]
-    async fn test_scheduler_scheduling_no_indexer() {
-        quickwit_common::setup_logging_for_tests();
-        let transport = ChannelTransport::default();
-        let cluster = create_cluster_for_test(Vec::new(), &["control_plane"], &transport, true)
-            .await
-            .unwrap();
-        let universe = Universe::with_accelerated_time();
-        let (indexing_service_inboxes, scheduler_handler) =
-            start_scheduler(cluster.clone(), &[], &universe).await;
-        assert_eq!(indexing_service_inboxes.len(), 0);
-
-        // No indexer.
-        universe.sleep(CONTROL_PLAN_LOOP_INTERVAL).await;
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 0);
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 0);
-        assert!(scheduler_state.last_applied_physical_plan.is_none());
-
-        // Wait REFRESH_PLAN_LOOP_INTERVAL * 2, as there is no indexer, we should observe no
-        // scheduling.
-        universe.sleep(REFRESH_PLAN_LOOP_INTERVAL * 2).await;
-        let scheduler_state = scheduler_handler.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 0);
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 0);
-        assert!(scheduler_state.last_applied_physical_plan.is_none());
-        universe.assert_quit().await;
-    }
-
-    #[tokio::test]
-    async fn test_scheduler_scheduling_multiple_indexers() {
-        quickwit_common::setup_logging_for_tests();
-        let transport = ChannelTransport::default();
-        let cluster = create_cluster_for_test(Vec::new(), &["control_plane"], &transport, true)
-            .await
-            .unwrap();
-        let cluster_indexer_1 = create_cluster_for_test(
-            vec![cluster.gossip_advertise_addr().to_string()],
-            &["indexer"],
-            &transport,
-            true,
-        )
-        .await
-        .unwrap();
-        let cluster_indexer_2 = create_cluster_for_test(
-            vec![cluster.gossip_advertise_addr().to_string()],
-            &["indexer"],
-            &transport,
-            true,
-        )
-        .await
-        .unwrap();
-        let universe = Universe::new();
-        let (indexing_service_inboxes, scheduler_handler) = start_scheduler(
-            cluster.clone(),
-            &[&cluster_indexer_1, &cluster_indexer_2],
-            &universe,
-        )
-        .await;
-        let indexing_service_inbox_1 = indexing_service_inboxes[0].clone();
-        let indexing_service_inbox_2 = indexing_service_inboxes[1].clone();
-        let scheduler_handler_arc = Arc::new(scheduler_handler);
-
-        // No indexer.
-        let scheduler_state = scheduler_handler_arc.process_pending_and_observe().await;
-        let indexing_service_inbox_messages =
-            indexing_service_inbox_1.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 0);
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 0);
-        assert!(scheduler_state.last_applied_physical_plan.is_none());
-        assert_eq!(indexing_service_inbox_messages.len(), 0);
-
-        cluster
-            .wait_for_ready_members(
-                |members| {
-                    members
-                        .iter()
-                        .any(|member| member.enabled_services.contains(&QuickwitService::Indexer))
-                },
-                Duration::from_secs(5),
-            )
-            .await
-            .unwrap();
-
-        // Wait for chitchat update, sheduler will detect new indexers and schedule a plan.
-        wait_until_predicate(
-            || {
-                let scheduler_handler_arc_clone = scheduler_handler_arc.clone();
-                async move {
-                    let scheduler_state = scheduler_handler_arc_clone
-                        .process_pending_and_observe()
-                        .await;
-                    scheduler_state.num_schedule_indexing_plan == 1
-                }
-            },
-            CONTROL_PLAN_LOOP_INTERVAL * 4,
-            Duration::from_millis(100),
-        )
-        .await
-        .unwrap();
-        let scheduler_state = scheduler_handler_arc.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_applied_physical_indexing_plan, 1);
-        let indexing_service_inbox_messages_1 =
-            indexing_service_inbox_1.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        let indexing_service_inbox_messages_2 =
-            indexing_service_inbox_2.drain_for_test_typed::<ApplyIndexingPlanRequest>();
-        assert_eq!(indexing_service_inbox_messages_1.len(), 1);
-        assert_eq!(indexing_service_inbox_messages_2.len(), 1);
-        cluster_indexer_1
-            .update_self_node_indexing_tasks(&indexing_service_inbox_messages_1[0].indexing_tasks)
-            .await
-            .unwrap();
-        cluster_indexer_2
-            .update_self_node_indexing_tasks(&indexing_service_inbox_messages_2[0].indexing_tasks)
-            .await
-            .unwrap();
-
-        // Wait 2 CONTROL_PLAN_LOOP_INTERVAL again and check the scheduler will not apply the plan
-        // several times.
-        universe.sleep(CONTROL_PLAN_LOOP_INTERVAL * 2).await;
-        let scheduler_state = scheduler_handler_arc.process_pending_and_observe().await;
-        assert_eq!(scheduler_state.num_schedule_indexing_plan, 1);
-
-        // Shutdown cluster and wait until the new scheduling.
-        cluster_indexer_2.shutdown().await;
-
-        cluster
-            .wait_for_ready_members(
-                |members| {
-                    members
-                        .iter()
-                        .filter(|member| {
-                            member.enabled_services.contains(&QuickwitService::Indexer)
-                        })
-                        .count()
-                        == 1
-                },
-                Duration::from_secs(5),
-            )
-            .await
-            .unwrap();
-
-        wait_until_predicate(
-            || {
-                let scheduler_handler_arc_clone = scheduler_handler_arc.clone();
-                async move {
-                    let scheduler_state = scheduler_handler_arc_clone
-                        .process_pending_and_observe()
-                        .await;
-                    scheduler_state.num_schedule_indexing_plan == 2
-                }
-            },
-            CONTROL_PLAN_LOOP_INTERVAL * 10,
-            Duration::from_millis(100),
-        )
-        .await
-        .unwrap();
-
-        universe.assert_quit().await;
-    }
+    use super::*;
 
     #[test]
     fn test_indexing_plans_diff() {
