+++ quickwit/quickwit-proto/src/codegen/quickwit/quickwit.control_plane.rs
@@ -1,14 +1,6 @@
 #[derive(serde::Serialize, serde::Deserialize, utoipa::ToSchema)]
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
-pub struct NotifyIndexChangeRequest {}
-#[derive(serde::Serialize, serde::Deserialize, utoipa::ToSchema)]
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct NotifyIndexChangeResponse {}
-#[derive(serde::Serialize, serde::Deserialize, utoipa::ToSchema)]
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
 pub struct GetOpenShardsRequest {
     #[prost(message, repeated, tag = "1")]
     pub subrequests: ::prost::alloc::vec::Vec<GetOpenShardsSubrequest>,
@@ -109,16 +101,6 @@ pub trait ControlPlaneService: std::fmt::Debug + dyn_clone::DynClone + Send + Sy
         &mut self,
         request: CloseShardsRequest,
     ) -> crate::control_plane::ControlPlaneResult<CloseShardsResponse>;
-    /// Notify the Control Plane that a change on an index occurred. The change
-    /// can be an index creation, deletion, or update that includes a source creation/deletion/num pipeline update.
-    /// Note(fmassot): it's not very clear for a user to know which change triggers a control plane notification.
-    /// This can be explicited in the attributes of `NotifyIndexChangeRequest` with an enum that describes the
-    /// type of change. The index ID and/or source ID could also be added.
-    /// However, these attributes will not be used by the Control Plane, at least at short term.
-    async fn notify_index_change(
-        &mut self,
-        request: NotifyIndexChangeRequest,
-    ) -> crate::control_plane::ControlPlaneResult<NotifyIndexChangeResponse>;
 }
 dyn_clone::clone_trait_object!(ControlPlaneService);
 #[cfg(any(test, feature = "testsuite"))]
@@ -231,12 +213,6 @@ impl ControlPlaneService for ControlPlaneServiceClient {
     ) -> crate::control_plane::ControlPlaneResult<CloseShardsResponse> {
         self.inner.close_shards(request).await
     }
-    async fn notify_index_change(
-        &mut self,
-        request: NotifyIndexChangeRequest,
-    ) -> crate::control_plane::ControlPlaneResult<NotifyIndexChangeResponse> {
-        self.inner.notify_index_change(request).await
-    }
 }
 #[cfg(any(test, feature = "testsuite"))]
 pub mod control_plane_service_mock {
@@ -298,12 +274,6 @@ pub mod control_plane_service_mock {
             request: super::CloseShardsRequest,
         ) -> crate::control_plane::ControlPlaneResult<super::CloseShardsResponse> {
             self.inner.lock().await.close_shards(request).await
-        }
-        async fn notify_index_change(
-            &mut self,
-            request: super::NotifyIndexChangeRequest,
-        ) -> crate::control_plane::ControlPlaneResult<super::NotifyIndexChangeResponse> {
-            self.inner.lock().await.notify_index_change(request).await
         }
     }
     impl From<MockControlPlaneService> for ControlPlaneServiceClient {
@@ -435,22 +405,6 @@ impl tower::Service<CloseShardsRequest> for Box<dyn ControlPlaneService> {
         Box::pin(fut)
     }
 }
-impl tower::Service<NotifyIndexChangeRequest> for Box<dyn ControlPlaneService> {
-    type Response = NotifyIndexChangeResponse;
-    type Error = crate::control_plane::ControlPlaneError;
-    type Future = BoxFuture<Self::Response, Self::Error>;
-    fn poll_ready(
-        &mut self,
-        _cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Result<(), Self::Error>> {
-        std::task::Poll::Ready(Ok(()))
-    }
-    fn call(&mut self, request: NotifyIndexChangeRequest) -> Self::Future {
-        let mut svc = self.clone();
-        let fut = async move { svc.notify_index_change(request).await };
-        Box::pin(fut)
-    }
-}
 /// A tower block is a set of towers. Each tower is stack of layers (middlewares) that are applied to a service.
 #[derive(Debug)]
 struct ControlPlaneServiceTowerBlock {
@@ -489,11 +443,6 @@ struct ControlPlaneServiceTowerBlock {
         CloseShardsResponse,
         crate::control_plane::ControlPlaneError,
     >,
-    notify_index_change_svc: quickwit_common::tower::BoxService<
-        NotifyIndexChangeRequest,
-        NotifyIndexChangeResponse,
-        crate::control_plane::ControlPlaneError,
-    >,
 }
 impl Clone for ControlPlaneServiceTowerBlock {
     fn clone(&self) -> Self {
@@ -505,7 +454,6 @@ impl Clone for ControlPlaneServiceTowerBlock {
             delete_source_svc: self.delete_source_svc.clone(),
             get_open_shards_svc: self.get_open_shards_svc.clone(),
             close_shards_svc: self.close_shards_svc.clone(),
-            notify_index_change_svc: self.notify_index_change_svc.clone(),
         }
     }
 }
@@ -554,12 +502,6 @@ impl ControlPlaneService for ControlPlaneServiceTowerBlock {
         request: CloseShardsRequest,
     ) -> crate::control_plane::ControlPlaneResult<CloseShardsResponse> {
         self.close_shards_svc.ready().await?.call(request).await
-    }
-    async fn notify_index_change(
-        &mut self,
-        request: NotifyIndexChangeRequest,
-    ) -> crate::control_plane::ControlPlaneResult<NotifyIndexChangeResponse> {
-        self.notify_index_change_svc.ready().await?.call(request).await
     }
 }
 #[derive(Debug, Default)]
@@ -627,15 +569,6 @@ pub struct ControlPlaneServiceTowerBlockBuilder {
             crate::control_plane::ControlPlaneError,
         >,
     >,
-    #[allow(clippy::type_complexity)]
-    notify_index_change_layer: Option<
-        quickwit_common::tower::BoxLayer<
-            Box<dyn ControlPlaneService>,
-            NotifyIndexChangeRequest,
-            NotifyIndexChangeResponse,
-            crate::control_plane::ControlPlaneError,
-        >,
-    >,
 }
 impl ControlPlaneServiceTowerBlockBuilder {
     pub fn shared_layer<L>(mut self, layer: L) -> Self
@@ -693,12 +626,6 @@ impl ControlPlaneServiceTowerBlockBuilder {
                 Error = crate::control_plane::ControlPlaneError,
             > + Clone + Send + Sync + 'static,
         <L::Service as tower::Service<CloseShardsRequest>>::Future: Send + 'static,
-        L::Service: tower::Service<
-                NotifyIndexChangeRequest,
-                Response = NotifyIndexChangeResponse,
-                Error = crate::control_plane::ControlPlaneError,
-            > + Clone + Send + Sync + 'static,
-        <L::Service as tower::Service<NotifyIndexChangeRequest>>::Future: Send + 'static,
     {
         self
             .create_index_layer = Some(
@@ -724,14 +651,7 @@ impl ControlPlaneServiceTowerBlockBuilder {
             .get_open_shards_layer = Some(
             quickwit_common::tower::BoxLayer::new(layer.clone()),
         );
-        self
-            .close_shards_layer = Some(
-            quickwit_common::tower::BoxLayer::new(layer.clone()),
-        );
-        self
-            .notify_index_change_layer = Some(
-            quickwit_common::tower::BoxLayer::new(layer),
-        );
+        self.close_shards_layer = Some(quickwit_common::tower::BoxLayer::new(layer));
         self
     }
     pub fn create_index_layer<L>(mut self, layer: L) -> Self
@@ -835,22 +755,6 @@ impl ControlPlaneServiceTowerBlockBuilder {
         self.close_shards_layer = Some(quickwit_common::tower::BoxLayer::new(layer));
         self
     }
-    pub fn notify_index_change_layer<L>(mut self, layer: L) -> Self
-    where
-        L: tower::Layer<Box<dyn ControlPlaneService>> + Send + Sync + 'static,
-        L::Service: tower::Service<
-                NotifyIndexChangeRequest,
-                Response = NotifyIndexChangeResponse,
-                Error = crate::control_plane::ControlPlaneError,
-            > + Clone + Send + Sync + 'static,
-        <L::Service as tower::Service<NotifyIndexChangeRequest>>::Future: Send + 'static,
-    {
-        self
-            .notify_index_change_layer = Some(
-            quickwit_common::tower::BoxLayer::new(layer),
-        );
-        self
-    }
     pub fn build<T>(self, instance: T) -> ControlPlaneServiceClient
     where
         T: ControlPlaneService,
@@ -924,12 +828,6 @@ impl ControlPlaneServiceTowerBlockBuilder {
         } else {
             quickwit_common::tower::BoxService::new(boxed_instance.clone())
         };
-        let notify_index_change_svc = if let Some(layer) = self.notify_index_change_layer
-        {
-            layer.layer(boxed_instance.clone())
-        } else {
-            quickwit_common::tower::BoxService::new(boxed_instance.clone())
-        };
         let tower_block = ControlPlaneServiceTowerBlock {
             create_index_svc,
             delete_index_svc,
@@ -938,7 +836,6 @@ impl ControlPlaneServiceTowerBlockBuilder {
             delete_source_svc,
             get_open_shards_svc,
             close_shards_svc,
-            notify_index_change_svc,
         };
         ControlPlaneServiceClient::new(tower_block)
     }
@@ -1077,15 +974,6 @@ where
                 CloseShardsResponse,
                 crate::control_plane::ControlPlaneError,
             >,
-        >
-        + tower::Service<
-            NotifyIndexChangeRequest,
-            Response = NotifyIndexChangeResponse,
-            Error = crate::control_plane::ControlPlaneError,
-            Future = BoxFuture<
-                NotifyIndexChangeResponse,
-                crate::control_plane::ControlPlaneError,
-            >,
         >,
 {
     async fn create_index(
@@ -1130,12 +1018,6 @@ where
         &mut self,
         request: CloseShardsRequest,
     ) -> crate::control_plane::ControlPlaneResult<CloseShardsResponse> {
-        self.call(request).await
-    }
-    async fn notify_index_change(
-        &mut self,
-        request: NotifyIndexChangeRequest,
-    ) -> crate::control_plane::ControlPlaneResult<NotifyIndexChangeResponse> {
         self.call(request).await
     }
 }
@@ -1233,16 +1115,6 @@ where
             .map(|response| response.into_inner())
             .map_err(|error| error.into())
     }
-    async fn notify_index_change(
-        &mut self,
-        request: NotifyIndexChangeRequest,
-    ) -> crate::control_plane::ControlPlaneResult<NotifyIndexChangeResponse> {
-        self.inner
-            .notify_index_change(request)
-            .await
-            .map(|response| response.into_inner())
-            .map_err(|error| error.into())
-    }
 }
 #[derive(Debug)]
 pub struct ControlPlaneServiceGrpcServerAdapter {
@@ -1332,17 +1204,6 @@ for ControlPlaneServiceGrpcServerAdapter {
         self.inner
             .clone()
             .close_shards(request.into_inner())
-            .await
-            .map(tonic::Response::new)
-            .map_err(|error| error.into())
-    }
-    async fn notify_index_change(
-        &self,
-        request: tonic::Request<NotifyIndexChangeRequest>,
-    ) -> Result<tonic::Response<NotifyIndexChangeResponse>, tonic::Status> {
-        self.inner
-            .clone()
-            .notify_index_change(request.into_inner())
             .await
             .map(tonic::Response::new)
             .map_err(|error| error.into())
@@ -1656,42 +1517,6 @@ pub mod control_plane_service_grpc_client {
                 );
             self.inner.unary(req, path, codec).await
         }
-        /// Notify the Control Plane that a change on an index occurred. The change
-        /// can be an index creation, deletion, or update that includes a source creation/deletion/num pipeline update.
-        /// Note(fmassot): it's not very clear for a user to know which change triggers a control plane notification.
-        /// This can be explicited in the attributes of `NotifyIndexChangeRequest` with an enum that describes the
-        /// type of change. The index ID and/or source ID could also be added.
-        /// However, these attributes will not be used by the Control Plane, at least at short term.
-        pub async fn notify_index_change(
-            &mut self,
-            request: impl tonic::IntoRequest<super::NotifyIndexChangeRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::NotifyIndexChangeResponse>,
-            tonic::Status,
-        > {
-            self.inner
-                .ready()
-                .await
-                .map_err(|e| {
-                    tonic::Status::new(
-                        tonic::Code::Unknown,
-                        format!("Service was not ready: {}", e.into()),
-                    )
-                })?;
-            let codec = tonic::codec::ProstCodec::default();
-            let path = http::uri::PathAndQuery::from_static(
-                "/quickwit.control_plane.ControlPlaneService/NotifyIndexChange",
-            );
-            let mut req = request.into_request();
-            req.extensions_mut()
-                .insert(
-                    GrpcMethod::new(
-                        "quickwit.control_plane.ControlPlaneService",
-                        "NotifyIndexChange",
-                    ),
-                );
-            self.inner.unary(req, path, codec).await
-        }
     }
 }
 /// Generated server implementations.
@@ -1755,19 +1580,6 @@ pub mod control_plane_service_grpc_server {
             request: tonic::Request<super::CloseShardsRequest>,
         ) -> std::result::Result<
             tonic::Response<super::CloseShardsResponse>,
-            tonic::Status,
-        >;
-        /// Notify the Control Plane that a change on an index occurred. The change
-        /// can be an index creation, deletion, or update that includes a source creation/deletion/num pipeline update.
-        /// Note(fmassot): it's not very clear for a user to know which change triggers a control plane notification.
-        /// This can be explicited in the attributes of `NotifyIndexChangeRequest` with an enum that describes the
-        /// type of change. The index ID and/or source ID could also be added.
-        /// However, these attributes will not be used by the Control Plane, at least at short term.
-        async fn notify_index_change(
-            &self,
-            request: tonic::Request<super::NotifyIndexChangeRequest>,
-        ) -> std::result::Result<
-            tonic::Response<super::NotifyIndexChangeResponse>,
             tonic::Status,
         >;
     }
@@ -2171,52 +1983,6 @@ pub mod control_plane_service_grpc_server {
                     let fut = async move {
                         let inner = inner.0;
                         let method = CloseShardsSvc(inner);
-                        let codec = tonic::codec::ProstCodec::default();
-                        let mut grpc = tonic::server::Grpc::new(codec)
-                            .apply_compression_config(
-                                accept_compression_encodings,
-                                send_compression_encodings,
-                            )
-                            .apply_max_message_size_config(
-                                max_decoding_message_size,
-                                max_encoding_message_size,
-                            );
-                        let res = grpc.unary(method, req).await;
-                        Ok(res)
-                    };
-                    Box::pin(fut)
-                }
-                "/quickwit.control_plane.ControlPlaneService/NotifyIndexChange" => {
-                    #[allow(non_camel_case_types)]
-                    struct NotifyIndexChangeSvc<T: ControlPlaneServiceGrpc>(pub Arc<T>);
-                    impl<
-                        T: ControlPlaneServiceGrpc,
-                    > tonic::server::UnaryService<super::NotifyIndexChangeRequest>
-                    for NotifyIndexChangeSvc<T> {
-                        type Response = super::NotifyIndexChangeResponse;
-                        type Future = BoxFuture<
-                            tonic::Response<Self::Response>,
-                            tonic::Status,
-                        >;
-                        fn call(
-                            &mut self,
-                            request: tonic::Request<super::NotifyIndexChangeRequest>,
-                        ) -> Self::Future {
-                            let inner = Arc::clone(&self.0);
-                            let fut = async move {
-                                (*inner).notify_index_change(request).await
-                            };
-                            Box::pin(fut)
-                        }
-                    }
-                    let accept_compression_encodings = self.accept_compression_encodings;
-                    let send_compression_encodings = self.send_compression_encodings;
-                    let max_decoding_message_size = self.max_decoding_message_size;
-                    let max_encoding_message_size = self.max_encoding_message_size;
-                    let inner = self.inner.clone();
-                    let fut = async move {
-                        let inner = inner.0;
-                        let method = NotifyIndexChangeSvc(inner);
                         let codec = tonic::codec::ProstCodec::default();
                         let mut grpc = tonic::server::Grpc::new(codec)
                             .apply_compression_config(
